Add files via upload
1. Overview of Network Topologies

Star Topology: In this configuration, all devices are connected to a central switch or hub. This central node acts as a conduit to transmit messages. The primary advantage is that if one peripheral device fails, it doesn't affect the rest of the network. However, if the central hub fails, the entire network is compromised. 

Bus Topology: All devices share a common communication line or bus. It's cost-effective for small networks but can lead to data collisions and is less reliable as more devices are added.